+++ RELEASENOTES.md
@@ -66,6 +66,9 @@
     *   Change the default value for
         `MediaLibrarySession.Builder.setLibraryErrorReplicationMode` to non
         fatal.
+    *   Add a `Context` parameter to
+        `MediaButtonReceiver.onForegroundServiceStartNotAllowedException`
+        ([#2625](https://github.com/androidx/media/pull/2625)).
 *   UI:
     *   Add `ProgressStateWithTickInterval` class and the corresponding
         `rememberProgressStateWithTickInterval` Composable to